refactor: domain.user.model 패키지 구조 변경
- model 패키지를 entity, repository로 분할
